add JWT authentication

+++ src/main/java/org/thesis/controllers/CartController.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.annotation.*;
 import org.thesis.models.CartRequest;
 import org.thesis.models.SimpleCart;
 import org.thesis.services.cart.CartService;
+import org.thesis.services.user.UserService;
 
 @RestController
 @RequestMapping("/api/v1/cart")
@@ -13,23 +14,46 @@ public class CartController {
     @Autowired
     private CartService cartService;
 
+    @Autowired
+    private UserService userService;
+
     @PostMapping("/add")
-    public ResponseEntity<SimpleCart> addToCart(@RequestBody CartRequest cartRequest) {
-        return cartService.addToCart(cartRequest);
+    public ResponseEntity<SimpleCart> addToCart(@RequestBody CartRequest cartRequest,
+                                                @RequestHeader("Authorization") String token) {
+        if (userService.validateToken(token)) {
+            return cartService.addToCart(cartRequest);
+        } else {
+            return ResponseEntity.status(403).build();
+        }
     }
 
     @PostMapping("/remove")
-    public ResponseEntity<SimpleCart> removeFromCart(@RequestBody CartRequest cartRequest) {
-        return cartService.removeFromCart(cartRequest);
+    public ResponseEntity<SimpleCart> removeFromCart(@RequestBody CartRequest cartRequest,
+                                                     @RequestHeader("Authorization") String token) {
+        if (userService.validateToken(token)) {
+            return cartService.removeFromCart(cartRequest);
+        } else {
+            return ResponseEntity.status(403).build();
+        }
     }
 
     @GetMapping("/{user}")
-    public ResponseEntity<SimpleCart> getCart(@PathVariable String user) {
-        return cartService.getCart(user);
+    public ResponseEntity<SimpleCart> getCart(@PathVariable String user,
+                                              @RequestHeader("Authorization") String token) {
+        if (userService.validateToken(token)) {
+            return cartService.getCart(user);
+        } else {
+            return ResponseEntity.status(403).build();
+        }
     }
 
     @PostMapping("/remove/{user}")
-    public ResponseEntity<SimpleCart> emptyCart(@PathVariable String user) {
-        return cartService.emptyCart(user);
+    public ResponseEntity<SimpleCart> emptyCart(@PathVariable String user,
+                                                @RequestHeader("Authorization") String token) {
+        if (userService.validateToken(token)) {
+            return cartService.emptyCart(user);
+        } else {
+            return ResponseEntity.status(403).build();
+        }
     }
 }
